+++ edgy/utils/inspect.py
@@ -8,6 +8,7 @@ from typing import Any, NoReturn
 import sqlalchemy
 from loguru import logger
 from monkay import load
+from sqlalchemy.dialects import postgresql
 from sqlalchemy.sql import schema, sqltypes
 
 import edgy
@@ -16,7 +17,7 @@ from edgy import Database, run_sync
 # Mapping of SQLAlchemy generic types to their corresponding Edgy field types.
 # This dictionary is crucial for translating database schema types into
 # Edgy model field definitions during database inspection.
-SQL_GENERIC_TYPES = {
+SQL_GENERIC_TYPES: dict[type, str] = {
     sqltypes.BigInteger: "edgy.core.db.fields.BigIntegerField",
     sqltypes.Integer: "edgy.core.db.fields.IntegerField",
     sqltypes.JSON: "edgy.core.db.fields.JSONField",
@@ -152,7 +153,7 @@ class InspectDB:
             tables, _ = self.generate_table_information(metadata)
 
             # Write the generated Edgy model code to standard output.
-            for line in self.write_output(tables, str(database.url), schema=self.schema):
+            for line in self.write_output(tables, database, schema=self.schema):
                 sys.stdout.writelines(line)  # type: ignore
 
     def inspect(self) -> None:
@@ -237,7 +238,7 @@ class InspectDB:
 
     @classmethod
     def get_field_type(
-        self, column: sqlalchemy.Column, is_fk: bool = False
+        self, column: sqlalchemy.Column, database: edgy.Database, is_fk: bool = False
     ) -> tuple[str, dict[str, Any]]:
         """
         Determines the appropriate Edgy field type and its parameters for a given
@@ -245,6 +246,7 @@ class InspectDB:
 
         Args:
             column (sqlalchemy.Column): The SQLAlchemy column to convert.
+            database (databasez.core.database.Database): The database used for inspection.
             is_fk (bool, optional): A flag indicating if the column is part of a
                                     foreign key relationship. If `True`, it will
                                     return "ForeignKey" or "OneToOne". Defaults to `False`.
@@ -283,6 +285,13 @@ class InspectDB:
             field_type = "TextField"
 
         # Populate field_params based on the determined field_type and column properties.
+
+        if (
+            field_type == "JSONField"
+            and database.url.dialect.startswith("postgres")
+            and not isinstance(column.type.dialect_impl(database.engine.dialect), postgresql.JSONB)
+        ):
+            field_params["no_jsonb"] = True
 
         if field_type == "PGArrayField":
             # For PGArrayField, we need the specific item_type, not its generic form.
@@ -405,7 +414,7 @@ class InspectDB:
 
     @classmethod
     def write_output(
-        cls, table_details: list[Any], connection_string: str, schema: str | None = None
+        cls, table_details: list[Any], database: edgy.Database, schema: str | None = None
     ) -> NoReturn:
         """
         Generates and yields lines of Python code representing the Edgy models.
@@ -414,7 +423,7 @@ class InspectDB:
 
         Args:
             table_details (list[Any]): A list of dictionaries, each detailing a reflected table.
-            connection_string (str): The database connection string used for the `Database` object.
+            database (databasez.core.database.Database): The database used for inspection.
             schema (str | None, optional): The schema name if one was used during reflection.
                                            Defaults to `None`.
 
@@ -448,7 +457,7 @@ class InspectDB:
 
         yield "\n"
         yield "\n"
-        yield f"database = {DB_MODULE}.Database('{connection_string}')\n"
+        yield f"database = {DB_MODULE}.Database('{database.url}')\n"
         yield registry
 
         # Iterate through each table detail to generate its class definition.
@@ -476,7 +485,7 @@ class InspectDB:
                 attr_name = column.name  # The Python attribute name for the field.
 
                 # Get the Edgy field type and its initial parameters.
-                field_type, field_params = cls.get_field_type(column, is_fk)
+                field_type, field_params = cls.get_field_type(column, database, is_fk)
                 field_params["null"] = column.nullable  # Add nullability.
 
                 # Handle primary key.